+++ tests/test_utilities/test_in2csv.py
@@ -257,3 +257,41 @@ class TestIn2CSV(CSVKitTestCase, EmptyFileTests):
                 path = 'examples/sheets_%d.csv' % suffix
                 if os.path.exists(path):
                     os.remove(path)
+
+    def test_convert_xls_with_write_sheets_with_names(self):
+        try:
+            self.assertConverted('xls', 'examples/sheets.xls', 'examples/testxls_converted.csv',
+                                 ['--sheet', 'data', '--write-sheets', "ʤ,1", '--use-sheet-names'])
+            with open('examples/sheets_ʤ.csv', 'r') as f:
+                with open('examples/testxls_unicode_converted.csv', 'r') as g:
+                    self.assertEqual(f.read(), g.read())
+            with open('examples/sheets_data.csv', 'r') as f:
+                with open('examples/testxls_converted.csv', 'r') as g:
+                    self.assertEqual(f.read(), g.read())
+            self.assertFalse(os.path.exists('examples/sheets_0.csv'))
+            self.assertFalse(os.path.exists('examples/sheets_1.csv'))
+            self.assertFalse(os.path.exists('examples/sheets_2.csv'))
+        finally:
+            for suffix in ('ʤ', 'data'):
+                path = 'examples/sheets_%s.csv' % suffix
+                if os.path.exists(path):
+                    os.remove(path)
+
+    def test_convert_xlsx_with_write_sheets_with_names(self):
+        try:
+            self.assertConverted('xlsx', 'examples/sheets.xlsx', 'examples/testxlsx_noinference_converted.csv',
+                                 ['--no-inference', '--sheet', 'data', '--write-sheets', "ʤ,1", '--use-sheet-names'])
+            with open('examples/sheets_ʤ.csv', 'r') as f:
+                with open('examples/testxlsx_unicode_converted.csv', 'r') as g:
+                    self.assertEqual(f.read(), g.read())
+            with open('examples/sheets_data.csv', 'r') as f:
+                with open('examples/testxlsx_noinference_converted.csv', 'r') as g:
+                    self.assertEqual(f.read(), g.read())
+            self.assertFalse(os.path.exists('examples/sheets_0.csv'))
+            self.assertFalse(os.path.exists('examples/sheets_1.csv'))
+            self.assertFalse(os.path.exists('examples/sheets_2.csv'))
+        finally:
+            for suffix in ('ʤ', 'data'):
+                path = 'examples/sheets_%s.csv' % suffix
+                if os.path.exists(path):
+                    os.remove(path)